Fixed a nasty race-condition crasher which appears after #4
Essentially, if current coordinator’s root VC is shown, but that coordinator is going away (being stopped and thus deallocated) and some other coordinator will take ownership of that same VC (this happens if that’s UINavigationController), there’s a race condition where parentCoordinator value on that rootVC is a nil reference. Which causes crash when reading out that property, due to ASSIGN being used in objc_associated

Yikes. So glad to have solved this.